+++ packages/contracts/contracts/test/harnesses/TestBalanceManager.sol
@@ -25,8 +25,13 @@ contract TestBalanceManager is IHandler, BalanceManager {
     function handleDeposit(
         Deposit calldata deposit
     ) external override onlyTeller {
-        StealthAddress calldata depositAddr = deposit.depositAddr;
-        _handleRefundNote(deposit.encodedAsset, depositAddr, deposit.value);
+        EncodedAsset[] memory encodedAssets = new EncodedAsset[](1);
+        uint256[] memory values = new uint256[](1);
+
+        encodedAssets[0] = deposit.encodedAsset;
+        values[0] = deposit.value;
+
+        _handleRefundNotes(encodedAssets, values, deposit.depositAddr, 1);
     }
 
     function handleOperation(