+++ packages/sdk/src/SparseMerkleProver.ts
@@ -9,6 +9,7 @@ import {
   partition,
   range,
 } from "./utils/functional";
+import { ZERO_VALUE, DEPTH, ARITY } from "./primitives/treeConstants";
 
 // high level idea:
 // want to sync a local replica of the tree such that
@@ -61,20 +62,18 @@ interface UncommittedLeaf {
 type PathIndex = number;
 
 const SMT_DUMP_KEY = "SMT_DUMP";
-export const ARITY = 4;
-export const MAX_DEPTH = 16;
 
 // TODO: turn these into constants
 // ZERO_HASH[i] = root hash of empty merkle tree of depth i
-export const ZERO_HASHES = [0n];
-for (let i = 1; i <= MAX_DEPTH; i++) {
+export const ZERO_HASHES = [ZERO_VALUE];
+for (let i = 1; i <= DEPTH; i++) {
   ZERO_HASHES.push(poseidonBN(new Array(ARITY).fill(ZERO_HASHES[i - 1])));
 }
 
 const NO_CHILDREN = new Array(ARITY).fill(undefined);
 
 function zeroHashAtDepth(depth: number): bigint {
-  return ZERO_HASHES[MAX_DEPTH - depth - 1];
+  return ZERO_HASHES[DEPTH - depth - 1];
 }
 
 function emptyNode(depth: number): TreeNode {
@@ -91,7 +90,7 @@ export class SparseMerkleProver {
   constructor(kv: KVStore) {
     this.root = {
       children: [...NO_CHILDREN],
-      hash: ZERO_HASHES[MAX_DEPTH],
+      hash: ZERO_HASHES[DEPTH],
     };
     this.leaves = new Map();
     this.uncommittedLeaves = [];
@@ -108,10 +107,7 @@ export class SparseMerkleProver {
   }
 
   insert(index: number, leaf: bigint, include = true): void {
-    assertOrErr(
-      index < ARITY ** MAX_DEPTH,
-      `index must be < ${ARITY}^maxDepth`
-    );
+    assertOrErr(index < ARITY ** DEPTH, `index must be < ${ARITY}^maxDepth`);
     assertOrErr(index >= this._count, "index must be >= tree count");
 
     this.root = this.insertInner(this.root, [leaf], pathIndexReverse(index));
@@ -125,7 +121,7 @@ export class SparseMerkleProver {
 
   insertBatch(startIndex: number, leaves: bigint[], includes: boolean[]): void {
     assertOrErr(
-      startIndex + leaves.length < ARITY ** MAX_DEPTH,
+      startIndex + leaves.length < ARITY ** DEPTH,
       `index must be < ${ARITY}^maxDepth`
     );
     assertOrErr(startIndex >= this._count, "index must be >= tree count");
@@ -150,10 +146,7 @@ export class SparseMerkleProver {
   }
 
   insertUncommitted(index: number, leaf: bigint, include = true): void {
-    assertOrErr(
-      index < ARITY ** MAX_DEPTH,
-      `index must be < ${ARITY}^maxDepth`
-    );
+    assertOrErr(index < ARITY ** DEPTH, `index must be < ${ARITY}^maxDepth`);
     assertOrErr(index >= this._count, "index must be >= tree count");
     assertOrErr(
       this.uncommittedLeaves.length === 0 ||
@@ -170,7 +163,7 @@ export class SparseMerkleProver {
     includes: boolean[]
   ): void {
     assertOrErr(
-      startIndex + leaves.length < ARITY ** MAX_DEPTH,
+      startIndex + leaves.length < ARITY ** DEPTH,
       `index must be < ${ARITY}^maxDepth`
     );
     assertOrErr(startIndex >= this._count, "index must be >= tree count");
@@ -181,7 +174,8 @@ export class SparseMerkleProver {
     assertOrErr(
       this.uncommittedLeaves.length === 0 ||
         startIndex >
-          this.uncommittedLeaves[this.uncommittedLeaves.length - 1].index
+          this.uncommittedLeaves[this.uncommittedLeaves.length - 1].index,
+      "insertions must be monotonic in index"
     );
 
     this.uncommittedLeaves.push(
@@ -347,7 +341,7 @@ export class SparseMerkleProver {
 
   // returns number of leaves in the subtree that we can't prune
   private pruneHelper(root: TreeNode, depth: number, index = 0): number {
-    if (depth === MAX_DEPTH && this.cannotPruneLeaf(index)) {
+    if (depth === DEPTH && this.cannotPruneLeaf(index)) {
       return 1;
     }
 
@@ -379,7 +373,7 @@ export class SparseMerkleProver {
     pathMask: number,
     depth = 0
   ): [bigint[][], PathIndex[]] {
-    if (depth === MAX_DEPTH) {
+    if (depth === DEPTH) {
       return [[], []];
     }
 
@@ -410,7 +404,7 @@ export class SparseMerkleProver {
     if (leaves.length === 0) return root;
 
     // we're at the leaf
-    if (depth === MAX_DEPTH) {
+    if (depth === DEPTH) {
       return { hash: leaves.shift()!, children: [...NO_CHILDREN] };
     }
 
@@ -443,7 +437,7 @@ export class SparseMerkleProver {
   private fillLeft(root: TreeNode, leaves: bigint[], depth: number): TreeNode {
     if (leaves.length === 0) return root;
 
-    if (depth === MAX_DEPTH) {
+    if (depth === DEPTH) {
       return { hash: leaves.shift()!, children: [...NO_CHILDREN] };
     }